requirements.txt
# FastAPI Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.9.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
email-validator>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1,<4.2.0

# Firebase
firebase-admin>=6.3.0

# Google AI
google-genai>=1.0.0

# Google Cloud TTS
google-cloud-texttospeech>=2.14.1

# Google Cloud Translation
google-cloud-translate>=3.15.0

# Audio Processing
pydub>=0.25.1
gtts>=2.5.0
edge-tts>=6.1.0

# HTTP Client
httpx>=0.28.1

# PDF/PPTX Generation
reportlab>=4.0.8
python-pptx>=0.6.23

# QR Code
qrcode[pil]>=7.4.2

# WebSockets
websockets>=12.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3

# Utils
python-dotenv>=1.0.0
